+++ lib/client/user-preferences.ts
@@ -9,22 +9,22 @@ export interface Preferences {
   path?: string;
 }
 
-export const getPreferences = async (): Promise<Preferences> => {
+const getPreferences = async (): Promise<Preferences> => {
   let errorMessage = "";
 
   const userHome = getHome() || "";
   const settings3 = join(
     userHome,
-    "/Library/Preferences/com.runningwithcrayons.Alfred-Preferences-3.plist",
+    "/Library/Preferences/com.runningwithcrayons.Alfred-Preferences-3.plist"
   );
   const settings = join(
     userHome,
-    "/Library/Preferences/com.runningwithcrayons.Alfred-Preferences.plist",
+    "/Library/Preferences/com.runningwithcrayons.Alfred-Preferences.plist"
   );
 
   const prefsJsonPath = join(
     userHome,
-    "/Library/Application Support/Alfred/prefs.json",
+    "/Library/Application Support/Alfred/prefs.json"
   );
 
   // PREFS JSON
@@ -44,13 +44,14 @@ export const getPreferences = async (): Promise<Preferences> => {
     errorMessage = `Alfred preferences not found at location ${prefsJsonPath}`;
   }
 
-  // PREFS PLIST
+  // PREFS PLIST Alfred 3
 
   try {
     const bplistFile = await Deno.readFile(settings3);
-    const { syncfolder = "~/Library/Application Support/Alfred 3" } = parseBuffer(bplistFile) as {
-      syncfolder?: string;
-    };
+    const { syncfolder = "~/Library/Application Support/Alfred 3" } =
+      parseBuffer(bplistFile) as {
+        syncfolder?: string;
+      };
     const prefsPath = join(syncfolder, "Alfred.alfredpreferences");
 
     return {
@@ -63,10 +64,12 @@ export const getPreferences = async (): Promise<Preferences> => {
     }
   }
 
+  // PREFS PLIST Alfred >3
+
   try {
     const bplistFile = await Deno.readFile(settings);
     const { syncfolder = "~/Library/Application Support/Alfred" } = parseBuffer(
-      bplistFile,
+      bplistFile
     ) as { syncfolder?: string };
     const prefsPath = join(syncfolder, "Alfred.alfredpreferences");
